Airflow/requirements.txt
faker
reportlab
Pillow
pymupdf
numpy<2.0.0
opencv-python
pdf2image
mysql-connector-python
openai
loguru
requests
pytesseract
scikit-learn
python-doctr[torch]
pymysql
apache-airflow-providers-amazon
boto3